+++ CHANGELOG.md
@@ -5,6 +5,11 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [2.1.0] - 2023-05-08
+### Changed
+- [@mjkuranda](https://github.com/mjkuranda): New color converting system.
+- [@mjkuranda](https://github.com/mjkuranda): Documentation for converters and `PaletteColor`.
+
 ## [2.0.0] - 2023-05-08
 ### Added
 - [@mjkuranda](https://github.com/mjkuranda): Permitting on conversion for images in `input` directory.